[EORL-11] 노티피케이션 테이블 생성 및 등록 쿼리 추가

+++ src/main/java/com/eorl/domain/member/member/MemberUpdateForm.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 public class MemberUpdateForm {
 
     @NotNull(message="memberId는 필수입니다.")
-    private int memberId;
+    private Long memberId;
     @NotNull(message="name은 필수입니다.")
     private String name;
 
@@ -26,7 +26,7 @@ public class MemberUpdateForm {
 
     private LocalDateTime authenticationDatetime;
 
-    public MemberUpdateForm(int memberId, String name, String password, String phoneNumber,
+    public MemberUpdateForm(Long memberId, String name, String password, String phoneNumber,
             String emailAddress) {
         this.memberId = memberId;
         this.name = name;
